+++ src/samara_variables.h
@@ -336,7 +336,7 @@ public:
 
 
 	//Sorghum Mortality por
-    std::array<int,5> tabCstr = {0,0,0,0,0};
+    std::array<double,5> tabCstr = {0,0,0,0,0};
 //    double tabCstr[6] = {0,0,0,0,0,0};
 	int tabCstrIndiceCourant = 0;
 	int NbJourCompte = 0;
